Changed: Objtypes are now from 2 up to max_objtype from pol.cfg 	Note:	 if objtypes <= pol.cfg max_tile_id and no graphic defined in itemdesc Pol uses objtype as graphic. 		 Multis and custom items should all be moved to > pol.cfg max_tile_id

+++ trunk/pol-core/pol/multi/boat.cpp
@@ -94,7 +94,7 @@ struct Component {
     const char* name;
     bool create;
     UBoat::BOAT_COMPONENT component_idx_;
-    unsigned short objtype_;
+    unsigned int objtype_;
 } boat_components[BoatShape::BOATSHAPE_COMPONENT__COUNT] = {
     { "Tillerman",                      true ,  UBoat::COMPONENT_TILLERMAN, EXTOBJ_TILLERMAN },
     { "PortGangplankExtended",          false,  UBoat::COMPONENT_PORT_PLANK, EXTOBJ_PORT_PLANK },
